Fix rendering of opacity when used on repeated elements
Opacity, visible and shadows on repeated elements work by replacing the
root element with a new element (opacity, visible clip, etc.) and making
the old one a child.

A tree like this

```
l := VerticalLayout {
    for i in 1: Rectangle {
        background: green;
        opacity: 0.2;
    }
}
```

is lowered like so (pseudo):

```
l := VerticalLayout {
    for i in 1: Opacity {
        opacity <=> r.opacity;
        r := Rectangle {
            background: green;
            opacity: 0.2;
            y: layout_cache[i].y;
            width: l.width;
            height: layout_cache[i].height;
        }
    }
}
```

and when rendering, it's important that the opacity of the Opacity
element is applied on the item renderer before rendering the Rectangle
r.

The Opacity element has no width/height because the default geometry
pass won't apply that to children of a layout.

Without a geometry, the item rendere will not call render() (no
intersection with the current clip), and thus the opacity is not
applied.

The shadow lowering pass handles this correctly, by moving the geometry
property bindings to the new root and letting the default geometry pass
apply 100% defaults (in the above "r" is not a layout child).

This patch moves this logic into the common
inject_element_as_repeated_element, so that we end up with a lowering
like this (after the default geometry pass):

```
l := VerticalLayout {
    for i in 1: op := Opacity {
        opacity <=> r.opacity;
        width: l.width;
        height: layout_cache[i].height;
        y: layout_cache[i].y;
        r := Rectangle {
            background: green;
            opacity: 0.2;
            width: op.width;
            height: op.height;
        }
    }
}
```