Set the rpath to Qt for Rust binaries
For C++ apps we apply rpath to the cdylib implicitly via qttypes, but
link flags for rust binaries are not transitively propagated but require
explicit opt-in by the top-level crate. However we want that
convenience, to avoid Rust apps having to deal with Qt backend specific
code in their build.rs. Therefore we use the _DEP mechanism to get the
qt library path from the qttypes crate, forward it to the default
backend crate and write it into a generic file in the qt backend's
build.rs, which is then picked up by sixtyfps-build.

In the future, we'll use the same mechanism to propagate link flags and
extend the linker serach path for the MCU build (such as the linker
script).

cc #566