Further trim the dist build of the online editor
Re-use the wasm interpreter from the web site when building for the web
site. It's built & copied in the CI at the same time, and we can
convince rollup to preserve the relative link for production builds,
while aliasing to the local directory for dev builds.

This brings down the editor to ~2.3MB on my machine.