Prospective fix for broken preview in doc snapshots
The doc preview in the snapshot tries to load https://sixtyfps.io/snapshots/master/wasm-interpreter/sixtyfps_wasm_interpreter.js
which doesn't exist. There's an index.js in that directory instead,
which is not a browser-friendly es module (it imports .wasm directly, as
one would do in a bundler). This might be a result of the webpack build, so
to be sure, build the wasm-interpreter by hand just after the editor instead of before.