Initial implementation of `access_item_rc` in the rust generator for llr
Otherwise this uses the trick of
InnnerCompo::item_tree()[self.tree_index] to retrieve the index of the
first children. This eliminates the tree_index_of_first_child member.

There are two test failures left:

One of them because accessing from within repeaters is not implemented
yet.

The second one because the self.tree_index of sub-components is not
calculated quite correctly yet when calling `init()`.