Add row data tracking to Rust sixtyfps::Model
Any row additional/removal marks the tracked row data as dirty. This
catches `someprop: some_mode[42];` being marked as dirty when a row
before 42 is inserted/removed (which makes sense). It'll also mark it as
dirty when a row after that is added/removed, but that keeps the code
simple for now.