Enable write access to models initialized from arrays in the interpreter
A model property such as

    property <[Foo]> model: [ ... ];

would create a VecModel in Rust and C++, which in turn can be written to
with an array index expression in .60. If the same .60 code was running
in the interpreter, a Value::Array would be created instead. We don't
support writing into Value::Array by index because we don't have
tracking in place for that, to update dependent bindings.

To fix this inconsistency, this patch changes the interpreter to
allocate a mutable VecModel as well.