Improve error handling when we fail to create a WebGL context in the browser
Bump the femtovg dependency to ensure we get the fix for error detection on the renderer side
and handle it in the GL backend side by rendering text to the 2D canvas, followed by a panic.

This was moved up to before the resize handler can be installed, as that would end up
clearing the canvas when resizing.

A message on the 2D canvas was chose over an alert message box, as the latter
is rather intrusive and would be a nightmare if many of them popped up
when viewing our docs with a browser
configured that way.