Test: enable the jpeg feature on the image crate
Since the cat picture is now in jpeg, we need to be able to decaode it
from the tests.
Luckily on the CI this works fine beacause we do a full cargo test that
build all the crates including the gl crate that has the jpeg format enaabled,
but if we try to compile the test with limited feature set, it wouldn't be available.