Fix application of popup window item background with the GL renderer
... and simplify the inline popup rendering code path across the renderers.

The GL backend failed to render the background of the `WindowItem` of a `PopupWindow`. This became visible with slint-ui/slint#1120 .

This change relaces the mandatory `draw_rectangle` in the item rendere with a mandatory `fill_rect`, which the former as well as the
centralized inline popup rendering can now use.

As an upside, this micro-optimizes away a property dependency to the background in
`draw_rectangle` for the QtItemRenderer
when the geometry was invalid, due to an early return.