a11y: Allow for (virtual) focus delegation at the a11y level
Have a property that allows to delegate the a11y-focus to one of
the children when a widget has the "real" focus. This makes e.g.
implementing the TabBar simple: The Tabbar widget has keyboard focus
and delegates to the individual tabs for a11y, so that those will be
read out.