Online editor: Add link to Slint language reference
This link works on the slint web site, but not when run locally. We do
not expect this to have users outside of the Slint site, so we are
willing to accept this limitation.

Please open a bug report if you indeed use the online editor elsewhere
and are bothered by the link.