Prospective fix for missing publish to vs code market place
We publish to OpenVSX first and if that succeeds then we publish to vs code market place.
The former sometimes times out, where the publishing actually succeeds but the GH
action step fails. Thus we never run the VS code market place step.

Swap the two steps as a starter and equip the OpenVSX step with continue-on-error: true
as at that point we have built the package, just publishing failed.