Fix compilation of generated Rust code when embedding fonts
The quote crate doesn't escape the characters for the character map entry.
This causes problems with some of the characters we chose by default, and in
general (if we ever do coverage based) will break. So for now just store
the char strait as u32.

There's an opportunity to use different character map encodings to cover the
common case that we don't actually need u32, but that's not for this patch.