Use prettyplease crate to do the formatting
This fixes compilation of generated Rust code when token stream contains
single-quoted semicolon.

When using glyph embedding, we generate a character map where each code
point is a literal char.  When the font contains a semicolon and we
generate an entry for that, we write ';' and the CodeFormatter would
think the semicolon is the end of a statement and produce a newline.
That breaks the build of the generated code.

Replace the hand-written formatter with the use of the prettyplease
crate.