Add hack to fix missing glyphs in the slide puzzle
The puzzle, printer, memory and launcher all register fonts. The very first font registered becomes the default font to use when
no font family is specified.

For the slide puzzle that means we end up detecting that we need in particular "↻✓", embed that in the slide puzzle's first registered font but
don't see it at run-time because the default font there is the one from the launcher.

This needs a different solution in the future, for now this gets the glyphs in place for the show.