GL backend: reduce memory allocations for text rendering
Fix a logic bug in the function to determine if we need to fall back to
other fonts or if the font we selected provides complete coverage.

In the common case where coverage is given, we start with some unknown
coverage (in old_uncovered_*) and end up with empty
remaining_required_*_coverage. That also happens to be less than
old_uncovered_*. Since that check came first, we always returned
"Improved", which meant we still built up the font fallback list, even
though we didn't need to. Fix the order of checks to avoid that.