Improve robustness of error reporting of the interpreter with dynamic scenes
With two-way bindings and dynamically created components it is possible
to end up with a binding to a property where the binding outlives the
property (see #1394).

With rust generated code, this means on binding evaluation, the upgrade
from self_weak to self_rc fails and panics.

With the interpreter we moved a VRef into the binding closure, which
ends up being a dangling pointer and causes memory corruption. This
change aligns the interpreter with the Rust way of creating closures by
creating the VRc first, moving a VWeak into the callback and binding
closures and upgrading that.