online_editor: Restructure the project
Use the typical src/ setup used in TS projects and split out the code
for the workers into src/worker so that this can get its own
configuration for the typescript compiler. This is necessary as the
web workers run in a more restricted environment as the rest of the
code and because `dom` and `webworker` can not be used together.