winit: Fix panic when shutting down event loop
If we (or the user) ends up calling hide() on a window after returning from the event loop, our
attempt at posting an event would panic in winit as we
try to create a new event loop instances.

Since Winit 0.27 a panic is raised when creating an event loop instance a second time.