online_editor: Implement custom "slint/load_file"
The LSP running in the web worker needs to access (virtual) files in the
editor. So it can send a request to the editor for files.

Handle such requests. Also make sure to provide the editor with the
correct URLs of open files so that the base URL to request additional
data with is correct.

Also do not error out when LSP and interpreter request the same file
concurrently.